Added: cmake cooperation, jpeg image reading (and many other). Removed: useless code (test, timer, c++11).
