+++ densecrf/refine_pascal/dense_inference.cpp
@@ -1,392 +1,410 @@
 /*
- * The code is modified from the NIPS demo code by Philipp Krähenbühl
+ * The code is modified from the NIPS demo code by Philipp Krahenbuhl
  *
- *
+ * Added: Support LoadMatFile and SaveMatFile. 
  *
  *
  */
 /*
-    Copyright (c) 2011, Philipp Krähenbühl
-    All rights reserved.
+        Copyright (c) 2011, Philipp Krahenbuhl
+        All rights reserved.
 
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions are met:
-        * Redistributions of source code must retain the above copyright
-        notice, this list of conditions and the following disclaimer.
-        * Redistributions in binary form must reproduce the above copyright
-        notice, this list of conditions and the following disclaimer in the
-        documentation and/or other materials provided with the distribution.
-        * Neither the name of the Stanford University nor the
-        names of its contributors may be used to endorse or promote products
-        derived from this software without specific prior written permission.
+        Redistribution and use in source and binary forms, with or without
+        modification, are permitted provided that the following conditions are met:
+                * Redistributions of source code must retain the above copyright
+                notice, this list of conditions and the following disclaimer.
+                * Redistributions in binary form must reproduce the above copyright
+                notice, this list of conditions and the following disclaimer in the
+                documentation and/or other materials provided with the distribution.
+                * Neither the name of the Stanford University nor the
+                names of its contributors may be used to endorse or promote products
+                derived from this software without specific prior written permission.
 
-    THIS SOFTWARE IS PROVIDED BY Philipp Krähenbühl ''AS IS'' AND ANY
-    EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-    WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL Philipp Krähenbühl BE LIABLE FOR ANY
-    DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-    (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-    LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-    ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+        THIS SOFTWARE IS PROVIDED BY Philipp Krahenbuhl ''AS IS'' AND ANY
+        EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+        WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+        DISCLAIMED. IN NO EVENT SHALL Philipp Krahenbuhl BE LIABLE FOR ANY
+        DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+        (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+        LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+        ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+        (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+        SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include <cstdio>
+#include "dense_inference.hpp"
 
-#include <iostream>
-#include <vector>
+#include <cassert>
 #include <cmath>
 #include <cstdlib>
-#include <string.h>
+#include <vector>
 #include <fstream>
+#include <utility>
+#include <ctime>
+
 #include <dirent.h>
 #include <fnmatch.h>
 
-#include "../libDenseCRF/densecrf.h"
-#include "../libDenseCRF/util.h"
-#include "dense_inference.h"
-#include "../util/Timer.h"
+#include "matio.h"
+#include "opencv2/highgui/highgui.hpp"
+#include "opencv2/imgproc/imgproc.hpp"
+#include "densecrf.h"
+
+
+template <typename Dtype> enum matio_classes matio_class_map();
+template <> enum matio_classes matio_class_map<float>() { return MAT_C_SINGLE; }
+template <> enum matio_classes matio_class_map<double>() { return MAT_C_DOUBLE; }
+template <> enum matio_classes matio_class_map<int>() { return MAT_C_INT32; }
+template <> enum matio_classes matio_class_map<unsigned int>() { return MAT_C_UINT32; }
+
 
 template <typename T>
-void LoadBinFile(std::string& fn, T*& data, 
-      int* row = NULL, int* col = NULL, int* channel = NULL);
+void LoadMatFile(const std::string& fileName, T*& data, int rows, int cols,
+        int& channels , bool do_ppm_format = false);
 
 template <typename T>
-void SaveBinFile(std::string& fn, T* data, 
-      int row = 1, int col = 1, int channel = 1);
+void LoadBinFile(const std::string& fileName, T*& data, int& rows, int& cols, int& channels);
 
 template <typename T>
-void LoadBinFile(std::string& fn, T*& data, 
-    int* row, int* col, int* channel) {
-  //data.clear();
+void SaveBinFile(const std::string& fileName, const T* data, int rows, int cols, int channels);
 
-  std::ifstream ifs(fn.c_str(), std::ios_base::in | std::ios_base::binary);
-
-  if(!ifs.is_open()) {
-    std::cerr << "Fail to open " << fn << std::endl;
-  }
-
-  int num_row, num_col, num_channel;
-
-  ifs.read((char*)&num_row, sizeof(int));
-  ifs.read((char*)&num_col, sizeof(int));
-  ifs.read((char*)&num_channel, sizeof(int));
-
-  int num_el;
-
-  num_el = num_row * num_col * num_channel;
-
-  //data.resize(num_el);
-  data = new T[num_el];
-
-  ifs.read((char*)&data[0], sizeof(T)*num_el);
-
-  ifs.close();
-
-  if(row!=NULL) {
-    *row = num_row;
-  }
-
-  if(col!=NULL) {
-    *col = num_col;
-  }
- 
-  if(channel != NULL) {
-    *channel = num_channel;
-  }
-
-}
 
 template <typename T>
-void SaveBinFile(std::string& fn, T* data, 
-    int row, int col, int channel) {
-  std::ofstream ofs(fn.c_str(), std::ios_base::out | std::ios_base::binary);
-
-  if(!ofs.is_open()) {
-    std::cerr << "Fail to open " << fn << std::endl;
-  }  
-
-  ofs.write((char*)&row, sizeof(int));
-  ofs.write((char*)&col, sizeof(int));
-  ofs.write((char*)&channel, sizeof(int));
-
-  int num_el;
-
-  num_el = row * col * channel;
-
-  ofs.write((char*)&data[0], sizeof(T)*num_el);
-
-  ofs.close();
-}
-
-
-void TraverseDirectory(const std::string& path, std::string& pattern, bool subdirectories, std::vector<std::string>& fileNames) {
-  DIR *dir, *tstdp;
-  struct dirent *dp;
-
-  //open the directory
-  if((dir  = opendir(path.c_str())) == NULL) {
-    std::cout << "Error opening " << path << std::endl;
-    return;
-  }
-
-  while ((dp = readdir(dir)) != NULL) {
-    tstdp=opendir(dp->d_name);
-		
-    if(tstdp) {
-      closedir(tstdp);
-      if(subdirectories) {
-	//TraverseDirectory(
-      }
-    } else {
-      if(fnmatch(pattern.c_str(), dp->d_name, 0)==0) {
-	//std::string tmp(path);	
-	//tmp.append("/").append(dp->d_name);
-	//fileNames.push_back(tmp);  //assume string ends with .bin
-
-	std::string tmp(dp->d_name);
-	fileNames.push_back(tmp.substr(0, tmp.length()-4));
-
-	//std::cout << fileNames.back() << std::endl;
-      }
+void LoadMatFile(const std::string& fileName, T*& data, int rows, int cols, int& channels, bool do_ppm_format) {
+    mat_t* matfp = Mat_Open(fileName.c_str(), MAT_ACC_RDONLY);
+    if (matfp == NULL) {
+        throw exception("Error opening MAT file: '" + fileName + "'.");
     }
-  }
 
-  closedir(dir);
-  return;
-}
-
-
-// Store the colors we read, so that we can write them again.
-int nColors = 0;
-int colors[255];
-
-unsigned int getColor( const unsigned char * c ){
-  return c[0] + 256*c[1] + 256*256*c[2];
-}
-void putColor( unsigned char * c, unsigned int cc ){
-  c[0] = cc&0xff; c[1] = (cc>>8)&0xff; c[2] = (cc>>16)&0xff;
-}
-// Produce a color image from a bunch of labels
-unsigned char * colorize( const short * map, int W, int H ){
-  unsigned char * r = new unsigned char[ W*H*3 ];
-  for( int k=0; k<W*H; k++ ){
-    int c = colors[ map[k] ];
-    putColor( r+3*k, c );
-  }
-  return r;
-}
-
-// Certainty that the groundtruth is correct
-const float GT_PROB = 0.5;
-
-// Simple classifier that is 50% certain that the annotation is correct
-float * classify( const unsigned char * im, int W, int H, int M ){
-  const float u_energy = -log( 1.0f / M );
-  const float n_energy = -log( (1.0f - GT_PROB) / (M-1) );
-  const float p_energy = -log( GT_PROB );
-  float * res = new float[W*H*M];
-  for( int k=0; k<W*H; k++ ){
-    // Map the color to a label
-    int c = getColor( im + 3*k );
-    int i;
-    for( i=0;i<nColors && c!=colors[i]; i++ );
-    if (c && i==nColors){
-      if (i<M)
-	colors[nColors++] = c;
-      else
-	c=0;
+    // Read data
+    matvar_t* matvar = Mat_VarReadInfo(matfp, "data");
+    if (matvar == NULL) {
+        Mat_Close(matfp);
+        throw exception("Field 'data' not present. In MAT file: '" + fileName + "'.");
     }
-		
-    // Set the energy
-    float * r = res + k*M;
-    if (c){
-      for( int j=0; j<M; j++ )
-	r[j] = n_energy;
-      r[i] = p_energy;
+
+    if (matvar->class_type != matio_class_map<T>()) {
+        Mat_VarFree(matvar);
+        Mat_Close(matfp);
+        throw exception("Field 'data' must be of the right class (single/double). In MAT file: '" + fileName + "'.");
     }
-    else{
-      for( int j=0; j<M; j++ )
-	r[j] = u_energy;
+    if ((4 <= matvar->rank) && (matvar->dims[3] != 1)) {
+        Mat_VarFree(matvar);
+        Mat_Close(matfp);
+        throw exception("Field 'data' cannot have ndims > 3. In MAT file: '" + fileName + "'.");
     }
-  }
-  return res;
-}
 
-struct InputData {
-  char* ImgDir;
-  char* FeatureDir;
-  char* SaveDir;
-  int MaxIterations;
-  float PosXStd;
-  float PosYStd;
-  float PosW;
-  float BilateralXStd;
-  float BilateralYStd;
-  float BilateralRStd;
-  float BilateralGStd;
-  float BilateralBStd;
-  float BilateralW;
-
-};
-
-int ParseInput(int argc, char** argv, struct InputData& OD) {
-  for(int k=1;k<argc;++k) {
-    if(::strcmp(argv[k], "-id")==0 && k+1!=argc) {
-      OD.ImgDir = argv[++k];
-    } else if(::strcmp(argv[k], "-fd")==0 && k+1!=argc) {
-      OD.FeatureDir = argv[++k];
-    } else if(::strcmp(argv[k], "-sd")==0 && k+1!=argc) {
-      OD.SaveDir = argv[++k];
-    } else if(::strcmp(argv[k], "-i")==0 && k+1!=argc) {
-      OD.MaxIterations = atoi(argv[++k]);
-    } else if(::strcmp(argv[k], "-px")==0 && k+1!=argc) {
-      OD.PosXStd = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-py")==0 && k+1!=argc) {
-      OD.PosYStd = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-pw")==0 && k+1!=argc) {
-      OD.PosW = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-bx")==0 && k+1!=argc) {
-      OD.BilateralXStd = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-by")==0 && k+1!=argc) {
-      OD.BilateralYStd = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-bw")==0 && k+1!=argc) {
-      OD.BilateralW = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-br")==0 && k+1!=argc) {
-      OD.BilateralRStd = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-bg")==0 && k+1!=argc) {
-      OD.BilateralGStd = atof(argv[++k]);
-    } else if(::strcmp(argv[k], "-bb")==0 && k+1!=argc) {
-      OD.BilateralBStd = atof(argv[++k]);
-    } 
-  }
-  return 0;
-}
-
-void ReshapeToMatlabFormat(short*& result, short* map, int img_row, int img_col) {
-  //row-order to column-order
-
-  int out_index, in_index;
-
-  for (int h = 0; h < img_row; ++h) {
-    for (int w = 0; w < img_col; ++w) {
-      out_index = w * img_row + h;
-      in_index  = h * img_col + w;
-      result[out_index] = map[in_index];
+    int fileSize = 1;
+    int dataSize = rows * cols;
+    for (int k = 0; k < matvar->rank; ++k) {
+        fileSize *= matvar->dims[k];
+        
+        if (1 < k) {
+            dataSize *= matvar->dims[k];
+        }
     }
-  }
-}
 
-void ComputeUnaryForCRF(float*& unary, float* feat, int feat_row, int feat_col, int feat_channel) {
-  int out_index, in_index;
+    assert(dataSize <= fileSize);
 
-  for (int h = 0; h < feat_row; ++h) {
-    for (int w = 0; w < feat_col; ++w) {
-      for (int c = 0; c < feat_channel; ++c) {
-	out_index = (h * feat_col + w) * feat_channel + c;
-	in_index  = (c * feat_col + w) * feat_row + h;
-	unary[out_index] = -log(feat[in_index]);
-      }
-    }
-  }
-}
-
-void OutputSetting(const InputData& inp) {
-  std::cout << "Input Parameters: " << std::endl;
-  std::cout << "ImgDir: "        << inp.ImgDir << std::endl;
-  std::cout << "FeatureDir: "    << inp.FeatureDir << std::endl;
-  std::cout << "SaveDir: "       << inp.SaveDir << std::endl;
-  std::cout << "MaxIterations: " << inp.MaxIterations << std::endl;
-  std::cout << "PosXStd: "  << inp.PosXStd << std::endl;
-  std::cout << "PosYStd: "  << inp.PosYStd << std::endl;
-  std::cout << "PosW: "     << inp.PosW    << std::endl;
-  std::cout << "Bi_X_Std: " << inp.BilateralXStd << std::endl;
-  std::cout << "Bi_Y_Std: " << inp.BilateralYStd << std::endl;
-  std::cout << "Bi_R_Std: " << inp.BilateralRStd << std::endl;
-  std::cout << "Bi_G_Std: " << inp.BilateralGStd << std::endl;
-  std::cout << "Bi_B_Std: " << inp.BilateralBStd << std::endl;
-  std::cout << "Bi_W: "     << inp.BilateralW    << std::endl;
-}
-
-int main( int argc, char* argv[]){
-  InputData inp;
-  inp.ImgDir = NULL;
-  inp.FeatureDir = NULL;
-  inp.SaveDir = NULL;
-  inp.MaxIterations = 10;
-  inp.PosXStd = 3;
-  inp.PosYStd = 3;
-  inp.PosW    = 3;
-  inp.BilateralXStd = 60;
-  inp.BilateralYStd = 60;
-  inp.BilateralRStd = 20;
-  inp.BilateralGStd = 20;
-  inp.BilateralBStd = 20;
-  inp.BilateralW    = 10;
-
-  ParseInput(argc, argv, inp);
-  OutputSetting(inp);
-
-  assert(inp.ImgDir != NULL && inp.FeatureDir != NULL && inp.SaveDir != NULL);
-
-  std::string pattern = "*.bin";
-  std::vector<std::string> feat_file_names;
-  std::string feat_folder(inp.FeatureDir);
-  TraverseDirectory(feat_folder, pattern, false, feat_file_names);
-  
-  float* feat;
-  float* unary;
-  unsigned char* img;
-  std::string fn;
-  int img_row, img_col; //, img_channel;
-  int feat_row, feat_col, feat_channel;
-
-  CPrecisionTimer CTmr;
-  CTmr.Start();
-  for (size_t i = 0; i < feat_file_names.size(); ++i) {
-    std::cout << "processing " << i << " (" << feat_file_names.size() << ")..." << std::endl;
-
-    fn = std::string(inp.ImgDir) + "/" + feat_file_names[i] + ".ppm";
-    img = readPPM(fn.c_str(), img_col, img_row);
-    //LoadBinFile(fn, img, &img_row, &img_col, &img_channel);
-
-    fn = std::string(inp.FeatureDir) + "/" + feat_file_names[i] + ".bin";
-    LoadBinFile(fn, feat, &feat_row, &feat_col, &feat_channel);
-
-    assert(img_row == feat_row && img_col == feat_col);
-
-    unary = new float[feat_row*feat_col*feat_channel];
-    ComputeUnaryForCRF(unary, feat, feat_row, feat_col, feat_channel);
-
-    // Setup the CRF model
-    DenseCRF2D crf(img_col, img_row, feat_channel);
-    // Specify the unary potential as an array of size W*H*(#classes)
-    // packing order: x0y0l0 x0y0l1 x0y0l2 .. x1y0l0 x1y0l1 ... (row-order)
-    crf.setUnaryEnergy(unary);
-    // add a color independent term (feature = pixel location 0..W-1, 0..H-1)
-    crf.addPairwiseGaussian(inp.PosXStd, inp.PosYStd, inp.PosW);
-
-    // add a color dependent term (feature = xyrgb)
-    crf.addPairwiseBilateral(inp.BilateralXStd, inp.BilateralYStd, inp.BilateralRStd, inp.BilateralGStd, inp.BilateralBStd, img, inp.BilateralW);
-	
-    // Do map inference
-    short* map = new short[img_row*img_col];
-    crf.map(inp.MaxIterations, map);
-
-    short* result = new short[img_row*img_col];
-    ReshapeToMatlabFormat(result, map, img_row, img_col);
-
-    // save results
-    fn = std::string(inp.SaveDir) + "/" + feat_file_names[i] + ".bin";
-    SaveBinFile(fn, result, feat_row, feat_col, 1);
+    T* fileData = new T[fileSize];
+    data = new T[dataSize];
     
-    // delete
-    delete[] result;
-    delete[] unary;
-    delete[] feat;
-    delete[] img;
-    delete[] map;
-  }
-  std::cout << "Time for inference: " << CTmr.Stop() << std::endl;
+    if (Mat_VarReadDataLinear(matfp, matvar, fileData, 0, 1, fileSize) != 0) {
+        delete[] data;
+        delete[] fileData;
+        Mat_VarFree(matvar);
+        Mat_Close(matfp);
+        throw exception("Error reading array 'data' from MAT file: '" + fileName + "'.");
+    }
+
+    // matvar->dims[0] : width
+    // matvar->dims[1] : height
+    const int channelSize = matvar->dims[0] * matvar->dims[1];
+    channels = static_cast<int>(matvar->dims[2]);
+
+    // extract from fileData
+    if (do_ppm_format) {
+        int out_offset = cols * channels;
+
+        for (int channel = 0; channel < channels; ++channel) {
+            for (int row = 0; row < rows; ++row) {
+                for (int column = 0; column < cols; ++column) {
+                    // perform transpose of fileData
+                    const int in_ind    = column + row * matvar->dims[0] + channel * channelSize;    
+                    const int out_ind = row * out_offset + column * channels + channel;
+                    // note the minus sign
+                    data[out_ind] = -fileData[in_ind]; 
+                }
+            }
+        }
+    } else {
+        int out_offset = rows * cols;
+
+        for (int channel = 0; channel < channels; ++channel) {
+            for (int row = 0; row < rows; ++row) {
+                for (int column = 0; column < cols; ++column) {
+                    const int in_ind    = row + column * matvar->dims[0] + channel * channelSize;
+                    const int out_ind = row + column * row + channel * out_offset; 
+                    data[out_ind] = -fileData[in_ind];        
+                }
+            }
+        }
+    }
+
+    Mat_VarFree(matvar);
+    Mat_Close(matfp);
+    delete[] fileData;
+}
+
+template <typename T>
+void LoadBinFile(const std::string& fileName, T*& data, int& rows, int& cols, int& channels) {
+    std::ifstream ifs(fileName.c_str(), std::ios_base::in | std::ios_base::binary);
+
+    if (ifs.is_open() == false) {
+        throw exception("Fail to open file: '" + fileName + "'.");
+    }
+
+    ifs.read((char*)&rows, sizeof(int));
+    ifs.read((char*)&cols, sizeof(int));
+    ifs.read((char*)&channels, sizeof(int));
+
+    const int num_el = rows * cols * channels;
+    data = new T[num_el];
+    ifs.read(reinterpret_cast<char*>(data), sizeof(T) * num_el);
+
+    ifs.close();
+}
+
+template <typename T>
+void SaveBinFile(const std::string& fileName, const T* data, int rows, int cols, int channels) {
+    std::ofstream ofs(fileName.c_str(), std::ios_base::out | std::ios_base::binary);
+
+    if (ofs.is_open() == false) {
+        throw exception("Fail to open file: '" + fileName + "'.");
+    }    
+
+    ofs.write((char*)&rows, sizeof(int));
+    ofs.write((char*)&cols, sizeof(int));
+    ofs.write((char*)&channels, sizeof(int));
+
+    int num_el = rows * cols * channels;
+    ofs.write(reinterpret_cast<const char*>(data), sizeof(T) * num_el);
+
+    ofs.close();
+}
+
+
+void ReshapeToMatlabFormat(short*& result, short* map, int rows, int cols) {
+    //row-order to column-order
+    for (int h = 0; h < rows; ++h) {
+        for (int w = 0; w < cols; ++w) {
+            const int out_index = w * rows + h;
+            const int in_index  = h * cols + w;
+            result[out_index] = map[in_index];
+        }
+    }
+}
+
+/*
+ Loads an image and coverts it to RGB(888) pixel format.
+ Returns false on fail, true on success.
+*/
+bool loadImage(const std::string& fileName, cv::Mat& image) {
+    image = cv::imread(fileName);
+    if (image.total() == 0) {
+        return false;
+    }
+    //convert to RGB pixel format
+    switch (image.type()) {
+    case CV_8UC3: 
+        cv::cvtColor(image, image, CV_BGR2RGB);
+        break;
+    case CV_8UC4:
+        cv::cvtColor(image, image, CV_BGRA2RGB);
+        break;
+    default: { /*none*/ }
+    }
+    return true;
+}
+
+
+std::ostream& operator<<(std::ostream& os, const InputData& inp) {
+    return os
+        << "Input Parameters: " << std::endl
+        << "imageDir:                     " << inp.imageDir << std::endl
+        << "featureDir:             " << inp.featureDir << std::endl
+        << "outputDir:                    " << inp.outputDir << std::endl
+        << "MaxIterations:        " << inp.maxIterations << std::endl
+    //    << "MaxImgSize:             " << inp.MaxImgSize << std::endl
+    //    << "NumClass:                 " << inp.NumClass << std::endl
+        << "posW:            " << inp.posW << std::endl
+        << "posXStd:     " << inp.posXStd << std::endl
+        << "posYStd:     " << inp.posYStd << std::endl
+        << "Bi_W:        " << inp.bilateralW << std::endl
+        << "Bi_X_Std:    " << inp.bilateralXStd << std::endl
+        << "Bi_Y_Std:    " << inp.bilateralYStd << std::endl
+        << "Bi_R_Std:    " << inp.bilateralRStd << std::endl
+        << "Bi_G_Std:    " << inp.bilateralGStd << std::endl
+        << "Bi_B_Std:    " << inp.bilateralBStd << std::endl;
+}
+
+InputData::InputData(int argc, char** argv)   
+{
+    init();
+    for(int k = 1; k < argc; ++k) {
+        if (std::strcmp(argv[k], "-id") == 0 && k + 1 != argc) {
+            imageDir = argv[++k];
+        } else if (std::strcmp(argv[k], "-fd")==0 && k+1!=argc) {
+            featureDir = argv[++k];
+        } else if (std::strcmp(argv[k], "-sd")==0 && k+1!=argc) {
+            outputDir = argv[++k];
+        } else if (std::strcmp(argv[k], "-i")==0 && k+1!=argc) {
+            maxIterations = atoi(argv[++k]);
+        } else if (std::strcmp(argv[k], "-px")==0 && k+1!=argc) {
+            posXStd = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-py")==0 && k+1!=argc) {
+            posYStd = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-pw")==0 && k+1!=argc) {
+            posW = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-bx")==0 && k+1!=argc) {
+            bilateralXStd = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-by")==0 && k+1!=argc) {
+            bilateralYStd = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-bw")==0 && k+1!=argc) {
+            bilateralW = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-br")==0 && k+1!=argc) {
+            bilateralRStd = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-bg")==0 && k+1!=argc) {
+            bilateralGStd = atof(argv[++k]);
+        } else if (std::strcmp(argv[k], "-bb")==0 && k+1!=argc) {
+            bilateralBStd = atof(argv[++k]);
+        } 
+    }
+}
+
+void InputData::init() {
+    maxIterations = 10;
+
+    posW    = 3;
+    posXStd = 3;
+    posYStd = 3;
+    bilateralW    = 5;
+    bilateralXStd = 70;
+    bilateralYStd = 70;
+    bilateralRStd = 5;
+    bilateralGStd = 5;
+    bilateralBStd = 5;
+}
+
+
+void listDirectory(const std::string& path, const std::string& pattern, bool stripExtension, std::vector<std::string>& fileNames) {
+    DIR* dir = opendir(path.c_str());
+    if (dir == NULL) {
+        throw exception("Error opening dir: '" + path + "'.");
+    }
+
+    dirent* entry = readdir(dir);
+    if (entry == NULL) {
+	    throw exception("Directory : '" + path + "' is empty!");
+    }
+
+    while (entry != NULL) {
+        DIR* testEntry = opendir(entry->d_name);
+        
+        if (testEntry == NULL) { // if not a directory
+            if (fnmatch(pattern.c_str(), entry->d_name, 0) == 0) {
+                if (stripExtension == true) {
+                    std::string tmp(entry->d_name);                    
+                    fileNames.emplace_back(std::move( tmp.substr(0, tmp.rfind(".")) ));
+                } else {
+                    fileNames.emplace_back(entry->d_name);
+                }
+            }
+        } else {
+            closedir(testEntry);
+        }
+	
+	    entry = readdir(dir);
+    }
+
+    closedir(dir);
+}
+
+void generateImageNames(const std::vector<std::string>& list, const std::string& stripPattern, std::vector<std::string>& out) {
+    for (auto it = list.cbegin(), iend = list.cend(); it != iend; ++it) {
+        size_t pos = (*it).find(stripPattern);
+        if (pos != std::string::npos) {
+            out.emplace_back(std::move( (*it).substr(0, pos) ));
+        }
+    }
+}
+
+int main(int argc, char* argv[]) {
+    InputData inp(argc, argv);
+    std::cout << inp;
+
+    assert(inp.imageDir.empty() == false && inp.featureDir.empty() == false && inp.outputDir.empty() == false);
+    
+    std::vector<std::string> featureNamesList;
+    listDirectory(inp.featureDir, "*.mat", true, featureNamesList);
+    
+    std::vector<std::string> imageNamesList;
+    generateImageNames(featureNamesList, "_blob_0", imageNamesList);
+
+    std::string fileName;
+    const bool do_ppm_format = true;
+    cv::Mat loadedImage;
+
+    const clock_t timeStart = clock();
+    for (size_t i = 0; i < featureNamesList.size(); ++i) {
+        if (i % 100 == 0) {
+            std::cout << "Processing " << i << " entry from " << featureNamesList.size() << " ..." << std::endl;
+        }
+
+        fileName = inp.imageDir + "/" + imageNamesList[i] + ".jpg";
+        
+        if (loadImage(fileName, loadedImage) == false) {
+            std::cerr << "Failed to open image: '" << fileName << "'. Skipping." << std::endl;
+            continue;
+        }
+        int featuresRows = loadedImage.rows;
+        int featuresColumns = loadedImage.cols;
+        unsigned char* img = loadedImage.data;
+
+        fileName = inp.featureDir + "/" + featureNamesList[i] + ".mat";
+
+        float* features = NULL;
+        int featuresChannels;
+        LoadMatFile(fileName, features, featuresRows, featuresColumns, featuresChannels, do_ppm_format);
+
+        // Setup the CRF model
+        DenseCRF2D crf(featuresColumns, featuresRows, featuresChannels);
+        // Specify the unary potential as an array of size W*H*(#classes)
+        // packing order: x0y0l0 x0y0l1 x0y0l2 .. x1y0l0 x1y0l1 ... (row-order)
+        crf.setUnaryEnergy(features);
+        // add a color independent term (feature = pixel location 0..W-1, 0..H-1)
+        crf.addPairwiseGaussian(inp.posXStd, inp.posYStd, inp.posW);
+
+        // add a color dependent term (feature = xyrgb)
+        crf.addPairwiseBilateral(inp.bilateralXStd, inp.bilateralYStd, inp.bilateralRStd, inp.bilateralGStd, inp.bilateralBStd, img, inp.bilateralW);
+    
+        // Do map inference
+        short* map = new short[featuresRows * featuresColumns];
+        crf.map(inp.maxIterations, map);
+
+        short* result = new short[featuresRows * featuresColumns];
+        ReshapeToMatlabFormat(result, map, featuresRows, featuresColumns);
+
+        // save results
+        fileName = inp.outputDir + "/" + imageNamesList[i] + ".bin";
+        SaveBinFile(fileName, result, featuresRows, featuresColumns, 1);
+        
+        delete[] result;
+        delete[] features;
+        delete[] map;
+    }
+
+    std::cout << "Time for inference: " << (double(clock() - timeStart) / CLOCKS_PER_SEC) << std::endl;
+
+    return 0;
 }
