--- densecrf/refine_pascal/dense_inference.h
@@ -1,12 +1,0 @@
-#ifndef _DENSE_INFERENCE_H
-#define _DENSE_INFERENCE_H
-
-#include <string>
-#include <vector>
-#include <fstream>
-#include <iostream>
-
-
-
-
-#endif
